+++ ta_lib/indicators/trend/src/smma.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn smma(source: &[f64], period: usize) -> Series<f64> {
+pub fn smma(source: &[f32], period: usize) -> Series<f32> {
     let source = Series::from(source);
 
     let smma = source.smma(period);
@@ -18,7 +18,7 @@ mod tests {
         let expected = vec![1.0, 1.333, 1.888, 2.592, 3.395];
         let epsilon = 0.001;
 
-        let result: Vec<f64> = smma(&source, 3).into();
+        let result: Vec<f32> = smma(&source, 3).into();
 
         for i in 0..source.len() {
             assert!(
