+++ ta_lib/indicators/volume/src/vo.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn vo(source: &[f64], short_period: usize, long_period: usize) -> Series<f64> {
+pub fn vo(source: &[f32], short_period: usize, long_period: usize) -> Series<f32> {
     let source = Series::from(source);
 
     let vo_short = source.ema(short_period);
@@ -21,7 +21,7 @@ mod tests {
         let expected = vec![0.0, 11.1111, 13.5802, 2.83224, -10.71604];
         let epsilon = 0.001;
 
-        let result: Vec<f64> = vo(&source, 2, 3).into();
+        let result: Vec<f32> = vo(&source, 2, 3).into();
 
         for i in 0..source.len() {
             assert!(
