+++ ta_lib/patterns/candlestick/src/counterattack.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn bullish(open: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bullish(open: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let close = Series::from(close);
 
@@ -10,7 +10,7 @@ pub fn bullish(open: &[f64], close: &[f64]) -> Series<bool> {
         & close.eq(&close.shift(1))
 }
 
-pub fn bearish(open: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bearish(open: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let close = Series::from(close);
