+++ ta_lib/patterns/candlestick/src/euphoria_extreme.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn bullish(open: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bullish(open: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let close = Series::from(close);
 
@@ -16,7 +16,7 @@ pub fn bullish(open: &[f64], close: &[f64]) -> Series<bool> {
         & body.shift(3).gt(&body.shift(4))
 }
 
-pub fn bearish(open: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bearish(open: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let close = Series::from(close);
