+++ ta_lib/patterns/candlestick/src/golden.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn bullish(open: &[f64], high: &[f64], low: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bullish(open: &[f32], high: &[f32], low: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let high = Series::from(high);
     let low = Series::from(low);
@@ -11,7 +11,7 @@ pub fn bullish(open: &[f64], high: &[f64], low: &[f64], close: &[f64]) -> Series
     low.lte(&open.shift(1)) & close.shift(1).gt(&golden_low) & close.shift(2).gt(&open.shift(2))
 }
 
-pub fn bearish(open: &[f64], high: &[f64], low: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bearish(open: &[f32], high: &[f32], low: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let high = Series::from(high);
     let low = Series::from(low);
